+++ web-app/src/App.js
@@ -1,7 +1,6 @@
 import React from 'react';
 import Navbar from './Components/Navbar/index';
 import Home from './Components/Home';
-import Search from './Components/Search';
 import SightDetails from './Components/Sights/Details';
 import SightForm from './Components/Sights/New';
 import Login from './Components/Login';
@@ -16,7 +15,6 @@ const App = () => (
       <Navbar />
       <Route exact path="/" component={Home} />
       <Route path="/home" component={Home} />
-      <Route path="/sights" component={Search} />
       <Route path="/sight/add" component={SightForm} />
       <Route path="/details/:id" component={SightDetails} />
       <Route path="/login" component={Login} />
